do not invoke_buy_resources_dialog() improperly
when can_topup is true but there are actually no resources needed. This is a corner case that can happen in TR due to network race condition. can_topup is left at "true" because there are actually no missing resources (on the client) and this leads to an improper call to invoke_buy_resources_dialog(), which TR does not support.