gameclient: when setting battle_outcome_dirty, also make sure to request a sync into battle_outcome_sync_marker
I think I figured out why battles sometimes didn't end at 100% base damage. There are a few places in the code that set session.battle_outcome_dirty = true, which is the flag to check for win/loss at the next combat tick. Each place is also supposed to set session.battle_outcome_sync_marker which delays that check until the server has caught up to all client network messages sent up to that point. But there was one code path that was setting session.battle_outcome_dirty to true without also setting the sync_marker, and therefore might be missing some of the last bits of damage. I'm not totally sure this will fix the bug, but it's more correct than it used to be.