gameclient: fix blocker targeting.
Units were previously flailing around attacking things not on the path to their original target because the code that looks for blockers incorrectly assumed that voxel_map_accel.objects_near_xy() only returned objects that actually block the given cell, when it actually can return other objects that are nearby but don't overlap that cell. Units were dutifully destroying all the blockers in a voxel instead of just the ones that block their path to the original target.