Disable the progression screens that show up before loot drops when there are 8 or fewer levels in a difficulty.
Normally they appear both before and after each loot drop. I did this because with the number of drops in Harvest, it was getting so that that screen was showing up after 75% of the levels and it felt like they were too frequent. Only 1/5 if this is actually necessary though.