fb_guest: change post-permission redirection to happen within the iframe only
Future Chrome versions will not allow iframed JavaScript to redirect the parent frame.

This is tricky because the following hit to ROOT will have a signed_request missing
the oauth_token. We have to pass the credentials returned by the Facebook JavaScript
SDK into proxyserver so it will complete the login.