FB login: don't accept oauth_tokens without checking for expiration first
There was a path - when spin_oauth_token was specified - where we would accept a login with an expired oauth_token.

We still required a valid signed_request, but those are long-lived, and not connected to the lifetime of the oauth_token.

Never noticed this before because it was rare to make a server side call using the client's oauth_token.