fix server-side handling of XSAPI requests when we can't find the active session
Most of the time, this case is caught by proxyserver, which correctly queues the payment and returns a "success" response to Xsolla. However, in rare timing scenarios, the request will make it into gameserver. This path was never tested and it looks like it never would have worked, because it falls through to self.handle_payment() with session being None.

The fix is to return immediately to Xsolla, since handle_payment() will be taken care of when the queued message is processed on next login.