[REVIEW] - Client/Verify - Weapon applies_aura array
I wanted the ability to have a spell apply more than one aura with different strengths and durations.  I made it so applies_aura can be a single string or an array, and the array can be a dictionary with aura_name, aura_duration, aura_strength, and aura_range, or a string with just the aura name.