[HOTFIX] - Client/Server - Mine level bug
I had a bug where the recipe level was leaking from one mine to another.

This caused crafting to fail if one mine tech was a higher level than the others.

I also revised the requirements not met error server message to identify the failed predicate if it's a level array instead of a single predicate.