[REVIEW] - Gameclient - Fixes current_item always null
Check MountedWeaponDialog.ondraw() at line 308. This was already setting current_item to null, then updating it based on the type of equipment the mounting object used.

At line 333, I added a line that passed current item as part of the sub-dialog's user_data.

Back in MountedWeaponDialog.set_recipe_display, it pulls the user data and uses that. It will never be undefined, because it is either null (which works fine) or the item set from the original API calls.

This works in DV, TR, and WSE, running the master branch and using player_db entries downloaded from the live game.  This has been tested with the various turret heads and with an empty turret emplacement.