[REVIEW] - Battalion repair res3 display
This updates dialogs.json to add 2-count and 3-count versions of the repair resource widgets.

SquadControlDialog is updated to determine if res3 is in play, and to show the correct widgets

Main.js is updated with a new abbreviated number display, since the numbers for repair can't fit in the new layout without abbreviation.