[REVIEW/DEPLOY] - Client - Add support for crafting_categories arrays
TR/DV's TOC12 needs to be a crafting, but TOC1-TOC11 needs to not be a crafter.  This required changing crafting_categories to an array with null entries for 1-11.

This further required adding support for checking if crafting_categories is an array of strings or an array with sub-arrays.

I have tested this in FS with the upcoming armed TOC, and also tested this in the TR/DV patch in progress.